+++ src/common/editor/area.designer.ts
@@ -6482,7 +6482,7 @@ export class AreaDesigner extends EditorBase {
                                                     <button id="btn-obj-material" class="btn-sm btn btn-default" style="width: 17px;min-width:17px;padding-left:4px;padding-right:4px;border-top-right-radius: 4px;border-bottom-right-radius: 4px;" type="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
                                                         <span class="caret" style="margin-left: -1px;"></span>
                                                     </button>
-                                                    <ul id="obj-material-list" style="max-height: 265px;" class="dropdown-menu pull-right" aria-labelledby="btn-obj-material" data-container="body">
+                                                    <ul id="obj-material-list" style="max-height: 200px;" class="dropdown-menu pull-right" aria-labelledby="btn-obj-material" data-container="body">
                                                     </ul>
                                                 </span>
                                             </div>
@@ -6749,7 +6749,7 @@ export class AreaDesigner extends EditorBase {
                             });
                             const qualities = `<div class="col-sm-12 form-group">
                             <label class="control-label" style="width: 100%;">Quality
-                                <select id="obj-quality" class="form-control selectpicker" data-style="btn-default btn-sm" data-width="100%">
+                                <select id="obj-quality" class="form-control selectpicker" data-size="8" data-style="btn-default btn-sm" data-width="100%">
                                     <option value='inferior'>Inferior</option>
                                     <option value='poor'>Poor</option>
                                     <option value='rough'>Rough</option>
@@ -6789,16 +6789,16 @@ export class AreaDesigner extends EditorBase {
                                                         <button id="btn-obj-limbs" class="btn-sm btn btn-default" style="width: 17px;min-width:17px;padding-left:4px;padding-right:4px;border-top-right-radius: 4px;border-bottom-right-radius: 4px;" type="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
                                                             <span class="caret" style="margin-left: -1px;"></span>
                                                         </button>
-                                                        <ul id="obj-limbs-list" style="max-height: 265px;" class="dropdown-menu pull-right" aria-labelledby="btn-obj-limbs" data-container="body">
+                                                        <ul id="obj-limbs-list" style="max-height: 200px;" class="dropdown-menu pull-right" aria-labelledby="btn-obj-limbs" data-container="body">
                                                             <li><a href="#">All limbs</a></li><li><a href="#">Overall</a></li><li><a href="#">Limb only</a></li><li><a href="#">Torso</a></li><li><a href="#">Head</a></li><li><a href="#">Left arm</a></li><li><a href="#">Right arm</a></li><li><a href="#">Left hand</a></li><li><a href="#">Right hand</a></li><li><a href="#">Left leg</a></li><li><a href="#">Right leg</a></li><li><a href="#">Left foot</a></li><li><a href="#">Right foot</a></li><li><a href="#">Right wing</a></li><li><a href="#">Left wing</a></li><li><a href="#">Left hoof</a></li><li><a href="#">Right hoof</a></li><li><a href="#">Tail</a></li><li><a href="#">Arms</a></li><li><a href="#">Legs</a></li><li><a href="#">Hands</a></li><li><a href="#">Feet</a></li><li><a href="#">Wings</a></li><li><a href="#">Hooves</a></li><li><a href="#">Lower body</a></li><li><a href="#">Core body</a></li><li><a href="#">Upper core</a></li><li><a href="#">Upper body</a></li><li><a href="#">Winged core</a></li><li><a href="#">Winged upper</a></li><li><a href="#">Upper trunk</a></li><li><a href="#">Lower trunk</a></li><li><a href="#">Trunk</a></li><li><a href="#">Winged trunk</a></li><li><a href="#">Full body</a></li><li><a href="#">Total body</a></li><li><a href="#">Winged body</a></li>
                                                         </ul>
                                                     </span>
                                                 </div>
                                             </label>
-                                        </div>${qualities.replace('col-sm-12', 'col-sm-6')}
+                                        </div>${qualities.replace('col-sm-12', 'col-sm-6').replace('data-size="8"', 'data-size="6"')}
                                         <div class="col-sm-6 form-group">
                                                 <label class="control-label" style="width: 100%;">Weapon type
-                                                    <select id="obj-wType" class="form-control selectpicker" data-style="btn-default btn-sm" data-width="100%">
+                                                    <select id="obj-wType" class="form-control selectpicker" data-size="6" data-style="btn-default btn-sm" data-width="100%">
                                                     <optgroup label="Axe"><option value="axe">Axe</option><option value="battle axe">Battle axe</option><option value="great axe">Great axe</option><option value="hand axe">Hand axe</option><option value="mattock">Mattock</option><option value="wood axe">Wood axe</option></optgroup><optgroup label="Blunt"><option value="club">Club</option><option value="hammer">Hammer</option><option value="mace">Mace</option><option value="maul">Maul</option><option value="morningstar">Morningstar</option><option value="spiked club">Spiked club</option><option value="warhammer">Warhammer</option></optgroup><optgroup label="Flail"><option value="ball and chain">Ball and chain</option><option value="chain">Chain</option><option value="flail">Flail</option><option value="whip">Whip</option></optgroup><optgroup label="Knife"><option value="dagger">Dagger</option><option value="dirk">Dirk</option><option value="knife">Knife</option><option value="kris">Kris</option><option value="stiletto">Stiletto</option><option value="tanto">Tanto</option></optgroup><optgroup label="Large sword"><option value="bastard sword">Bastard sword</option><option value="claymore">Claymore</option><option value="flamberge">Flamberge</option><option value="large sword">Large sword</option><option value="nodachi">Nodachi</option></optgroup><optgroup label="Melee"><option value="brass knuckles">Brass knuckles</option><option value="melee">Melee</option><option value="tekagi-shuko">Tekagi-shuko</option><option value="tekko">Tekko</option></optgroup><optgroup label="Miscellaneous"><option value="cord">Cord</option><option value="fan">Fan</option><option value="giant fan">Giant fan</option><option value="miscellaneous">Miscellaneous</option><option value="war fan">War fan</option></optgroup><optgroup label="Polearm"><option value="bardiche">Bardiche</option><option value="glaive">Glaive</option><option value="halberd">Halberd</option><option value="poleaxe">Poleaxe</option><option value="scythe">Scythe</option></optgroup><optgroup label="Small sword"><option value="broadsword">Broadsword</option><option value="katana">Katana</option><option value="long sword">Long sword</option><option value="rapier">Rapier</option><option value="scimitar">Scimitar</option><option value="short sword">Short sword</option><option value="small sword">Small sword</option><option value="wakizashi">Wakizashi</option></optgroup><optgroup label="Spear"><option value="arrow">Arrow</option><option value="javelin">Javelin</option><option value="lance">Lance</option><option value="long spear">Long spear</option><option value="pike">Pike</option><option value="pilum">Pilum</option><option value="short spear">Short spear</option><option value="spear">Spear</option><option value="trident">Trident</option></optgroup><optgroup label="Staff"><option value="battle staff">Battle staff</option><option value="bo">Bo</option><option value="quarterstaff">Quarterstaff</option><option value="staff">Staff</option><option value="wand">Wand</option><option value="warstaff">Warstaff</option></optgroup>
                                                     </select>
                                                 </label>
@@ -6977,13 +6977,13 @@ export class AreaDesigner extends EditorBase {
                                                     <button id="btn-obj-limbs" class="btn-sm btn btn-default" style="width: 17px;min-width:17px;padding-left:4px;padding-right:4px;border-top-right-radius: 4px;border-bottom-right-radius: 4px;" type="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
                                                         <span class="caret" style="margin-left: -1px;"></span>
                                                     </button>
-                                                    <ul id="obj-limbs-list" style="max-height: 265px;" class="dropdown-menu pull-right" aria-labelledby="btn-obj-limbs" data-container="body">
+                                                    <ul id="obj-limbs-list" style="max-height: 200px;" class="dropdown-menu pull-right" aria-labelledby="btn-obj-limbs" data-container="body">
                                                         <li><a href="#">All limbs</a></li><li><a href="#">Overall</a></li><li><a href="#">Limb only</a></li><li><a href="#">Torso</a></li><li><a href="#">Head</a></li><li><a href="#">Left arm</a></li><li><a href="#">Right arm</a></li><li><a href="#">Left hand</a></li><li><a href="#">Right hand</a></li><li><a href="#">Left leg</a></li><li><a href="#">Right leg</a></li><li><a href="#">Left foot</a></li><li><a href="#">Right foot</a></li><li><a href="#">Right wing</a></li><li><a href="#">Left wing</a></li><li><a href="#">Left hoof</a></li><li><a href="#">Right hoof</a></li><li><a href="#">Tail</a></li><li><a href="#">Arms</a></li><li><a href="#">Legs</a></li><li><a href="#">Hands</a></li><li><a href="#">Feet</a></li><li><a href="#">Wings</a></li><li><a href="#">Hooves</a></li><li><a href="#">Lower body</a></li><li><a href="#">Core body</a></li><li><a href="#">Upper core</a></li><li><a href="#">Upper body</a></li><li><a href="#">Winged core</a></li><li><a href="#">Winged upper</a></li><li><a href="#">Upper trunk</a></li><li><a href="#">Lower trunk</a></li><li><a href="#">Trunk</a></li><li><a href="#">Winged trunk</a></li><li><a href="#">Full body</a></li><li><a href="#">Total body</a></li><li><a href="#">Winged body</a></li>
                                                     </ul>
                                                 </span>
                                             </div>
                                         </label>
-                                    </div>${qualities}
+                                    </div>${qualities.replace('data-size="8"', 'data-size="6"')}
                                     <div class="form-group col-sm-6">
                                         <label class="control-label">
                                             Enchantment
@@ -7126,7 +7126,7 @@ export class AreaDesigner extends EditorBase {
                                         enterMoveNew: this.$enterMoveNew
                                     }), wizSkills, wizBonuses]);
                                     if (StdObjectType.armor_of_holding == ty) {
-                                        wiz.insertPages(1, new WizardPage({
+                                        wiz.insertPages(8, new WizardPage({
                                             id: 'obj-bag',
                                             title: 'Armor of holding properties',
                                             body: ` <div class="col-sm-6 form-group">
@@ -7330,7 +7330,7 @@ export class AreaDesigner extends EditorBase {
                                                     <button id="btn-obj-size" class="btn-sm btn btn-default" style="width: 17px;min-width:17px;padding-left:4px;padding-right:4px;border-top-right-radius: 4px;border-bottom-right-radius: 4px;" type="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
                                                         <span class="caret" style="margin-left: -1px;"></span>
                                                     </button>
-                                                    <ul id="obj-size-list" style="max-height: 265px;" class="dropdown-menu pull-right" aria-labelledby="btn-obj-size" data-container="body">
+                                                    <ul id="obj-size-list" style="max-height: 200px;" class="dropdown-menu pull-right" aria-labelledby="btn-obj-size" data-container="body">
                                                         <li><a href="#">Small</a></li>
                                                         <li><a href="#">Medium</a></li>
                                                         <li><a href="#">Large</a></li>
@@ -7371,7 +7371,7 @@ export class AreaDesigner extends EditorBase {
                                                     <button id="btn-obj-size" class="btn-sm btn btn-default" style="width: 17px;min-width:17px;padding-left:4px;padding-right:4px;border-top-right-radius: 4px;border-bottom-right-radius: 4px;" type="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
                                                         <span class="caret" style="margin-left: -1px;"></span>
                                                     </button>
-                                                    <ul id="obj-size-list" style="max-height: 265px;" class="dropdown-menu pull-right" aria-labelledby="btn-obj-size" data-container="body">
+                                                    <ul id="obj-size-list" style="max-height: 200px;" class="dropdown-menu pull-right" aria-labelledby="btn-obj-size" data-container="body">
                                                         <li><a href="#">Small</a></li>
                                                         <li><a href="#">Medium</a></li>
                                                         <li><a href="#">Large</a></li>
@@ -7414,11 +7414,11 @@ export class AreaDesigner extends EditorBase {
                                     </div>
                                     <div class="col-sm-12 form-group">
                                         <label class="control-label" style="width: 100%;">Weapon type
-                                            <select id="obj-wType" class="form-control selectpicker" data-style="btn-default btn-sm" data-width="100%">
+                                            <select id="obj-wType" class="form-control selectpicker" data-size="8" data-style="btn-default btn-sm" data-width="100%">
                                             <option value="">Default</option><optgroup label="Axe"><option value="axe">Axe</option><option value="battle axe">Battle axe</option><option value="great axe">Great axe</option><option value="hand axe">Hand axe</option><option value="mattock">Mattock</option><option value="wood axe">Wood axe</option></optgroup><optgroup label="Blunt"><option value="club">Club</option><option value="hammer">Hammer</option><option value="mace">Mace</option><option value="maul">Maul</option><option value="morningstar">Morningstar</option><option value="spiked club">Spiked club</option><option value="warhammer">Warhammer</option></optgroup><optgroup label="Flail"><option value="ball and chain">Ball and chain</option><option value="chain">Chain</option><option value="flail">Flail</option><option value="whip">Whip</option></optgroup><optgroup label="Knife"><option value="dagger">Dagger</option><option value="dirk">Dirk</option><option value="knife">Knife</option><option value="kris">Kris</option><option value="stiletto">Stiletto</option><option value="tanto">Tanto</option></optgroup><optgroup label="Large sword"><option value="bastard sword">Bastard sword</option><option value="claymore">Claymore</option><option value="flamberge">Flamberge</option><option value="large sword">Large sword</option><option value="nodachi">Nodachi</option></optgroup><optgroup label="Melee"><option value="brass knuckles">Brass knuckles</option><option value="melee">Melee</option><option value="tekagi-shuko">Tekagi-shuko</option><option value="tekko">Tekko</option></optgroup><optgroup label="Miscellaneous"><option value="cord">Cord</option><option value="fan">Fan</option><option value="giant fan">Giant fan</option><option value="miscellaneous">Miscellaneous</option><option value="war fan">War fan</option></optgroup><optgroup label="Polearm"><option value="bardiche">Bardiche</option><option value="glaive">Glaive</option><option value="halberd">Halberd</option><option value="poleaxe">Poleaxe</option><option value="scythe">Scythe</option></optgroup><optgroup label="Small sword"><option value="broadsword">Broadsword</option><option value="katana">Katana</option><option value="long sword">Long sword</option><option value="rapier">Rapier</option><option value="scimitar">Scimitar</option><option value="short sword">Short sword</option><option value="small sword">Small sword</option><option value="wakizashi">Wakizashi</option></optgroup><optgroup label="Spear"><option value="javelin">Javelin</option><option value="lance">Lance</option><option value="long spear">Long spear</option><option value="pike">Pike</option><option value="pilum">Pilum</option><option value="short spear">Short spear</option><option value="spear">Spear</option><option value="trident">Trident</option></optgroup><optgroup label="Staff"><option value="battle staff">Battle staff</option><option value="bo">Bo</option><option value="quarterstaff">Quarterstaff</option><option value="staff">Staff</option><option value="wand">Wand</option><option value="warstaff">Warstaff</option></optgroup>
                                             </select>
                                         </label>
-                                    </div>${qualities}
+                                    </div>${qualities.replace('data-size="8"', 'data-size="6"')}
                                     <div class="form-group col-sm-12">
                                         <label class="control-label">
                                             Enchantment
@@ -7512,7 +7512,7 @@ export class AreaDesigner extends EditorBase {
                                                     <button id="btn-obj-size" class="btn-sm btn btn-default" style="width: 17px;min-width:17px;padding-left:4px;padding-right:4px;border-top-right-radius: 4px;border-bottom-right-radius: 4px;" type="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
                                                         <span class="caret" style="margin-left: -1px;"></span>
                                                     </button>
-                                                    <ul id="obj-size-list" style="max-height: 265px;" class="dropdown-menu pull-right" aria-labelledby="btn-obj-size" data-container="body">
+                                                    <ul id="obj-size-list" style="max-height: 200px;" class="dropdown-menu pull-right" aria-labelledby="btn-obj-size" data-container="body">
                                                         <li><a href="#">Small</a></li>
                                                         <li><a href="#">Medium</a></li>
                                                         <li><a href="#">Large</a></li>
@@ -7566,7 +7566,7 @@ export class AreaDesigner extends EditorBase {
                                                 <button id="btn-obj-size" class="btn-sm btn btn-default" style="width: 17px;min-width:17px;padding-left:4px;padding-right:4px;border-top-right-radius: 4px;border-bottom-right-radius: 4px;" type="button" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
                                                     <span class="caret" style="margin-left: -1px;"></span>
                                                 </button>
-                                                <ul id="obj-preserved-list" style="max-height: 265px;" class="dropdown-menu pull-right" aria-labelledby="btn-obj-size" data-container="body">
+                                                <ul id="obj-preserved-list" style="max-height: 200px;" class="dropdown-menu pull-right" aria-labelledby="btn-obj-size" data-container="body">
                                                     <li><a href="#">Cooked</a></li>
                                                     <li><a href="#">Smoked</a></li>
                                                     <li><a href="#">Salted</a></li>